src/gcpUtils/requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
google-api-core==1.29.0
google-auth==1.30.1
google-cloud-bigquery==2.20.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-cloud-vision==2.3.1
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.38.0
idna==2.10
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
urllib3==1.26.5